requirements.txt
#logging
Flask[async]
python-dotenv
asyncio
aiohttp
aiofiles
flask-cors
Pillow

